[ONT-37] Put slot in base table row for now
